Extract resumption tests
